Struct rea_rs::audio_accessor::AudioAccessor

pub struct AudioAccessor<'a, T: KnowsProject, P: ProbablyMutable> { /* private fields */ }

Implementations

impl<'a, T: KnowsProject, P: ProbablyMutable> AudioAccessor<'a, T, P>

pub fn new(parent: &'a T, ptr: AudioAccessor) -> Self

pub fn has_state_changed(&self) -> bool

pub fn start(&self) -> Position

pub fn end(&self) -> Position

pub fn get_sample_block_raw(
&self,
start: SampleAmount,
samples_per_channel: u32,
n_channels: u8,
samplerate: u32
) -> ReaperResult<Option<Vec<f64>>>

Get buffer of samples with given absolute project position in samples.

Returned buffer is Vec<f64> of length samples_per_channel * n_channels: values placed sample by sample for each channel in a turn. E.g. [spl1_ch1, spl1_ch2, spl1_ch3, spl2_ch1, spl2_ch2, spl3_ch3]

start is amount of samples since audio accessor start time.

You can look at the routing scheme to see where AudioAccessor is connected in case of track or take parenting: https://wiki.cockos.com/wiki/images/5/50/SWS_loudness_analysis_signal_flow_chart.png

Example from SWS: https://github.com/reaper-oss/sws/blob/bcc8fbc96f30a943bd04fb8030b4a03ea1ff7557/Breeder/BR_Loudness.cpp#L1020-L1079

note

Samples converted back to seconds as persized as possible, but I still afraid a bit of this function being sample-accurate.

impl<'a, T: KnowsProject> AudioAccessor<'a, T, Mutable>

pub fn validate(&mut self) -> bool

Validates the current state of the audio accessor

– must ONLY call this from the main thread.

Returns true if the state changed.

pub fn update(&mut self)

Force the accessor to reload its state from the underlying track or media item take.