Struct rea_rs::envelope::Envelope

pub struct Envelope<'a, P: KnowsProject, T: ProbablyMutable> { /* private fields */ }

Can be either TrackEnvelope, or TakeEnvelope

Main points of struct construction are:

crate::Track::get_envelope

crate::Track::get_envelope_by_chunk (or guid)

crate::Track::get_envelope_by_name (I’m not sure it works.)

[crate::TrackSend::get_envelope]

[crate::FxParam::add_envelope]

[crate::Take::get_envelope]

Implementations

impl<'a, P: KnowsProject, T: ProbablyMutable> Envelope<'a, P, T>

pub fn new(ptr: TrackEnvelope, parent: &'a P) -> Self

pub fn guid(&self) -> GUID

pub fn get_automation_item(
&'a self,
index: usize
) -> Option<AutomationItem<'a, P, T>>

pub fn name(&self) -> String

pub fn state_chunk(&self) -> String

Full Envelope state, as it written in project file.

pub fn n_points(&self) -> usize

pub fn n_automation_items(&self) -> usize

pub fn get_point(&self, point_index: usize) -> ReaperStaticResult<EnvelopePoint>

pub fn get_point_by_time(
&self,
position: impl Into<Position>
) -> ReaperStaticResult<EnvelopePoint>

Get the last point before the given time.

pub fn evaluate(
&self,
position: Position,
samplerate: u32,
samples_requested: usize
) -> EnvelopeEvaluateResult

Get the effective envelope value at a given time position.

samples_requested is how long the caller expects until the next call to evaluate (often, the buffer block size).

See Envelope::scaling_mode.

pub fn format_value(&self, value: f64) -> String

Get value, as it written in GUI

pub fn get_info_value(&self, category: impl Into<String>) -> f64

Get info value by string key. Should not be used in 99% cases.

Is public only for the case of pointers information retrieval, if suddenly needed.

unimplemented categories:

P_TRACK : MediaTrack * : parent track pointer (if any)

P_DESTTRACK : MediaTrack * : destination track pointer, if on a send

P_ITEM : MediaItem * : parent item pointer (if any)

P_TAKE : MediaItem_Take * : parent take pointer (if any)

pub fn send_info(&self) -> Option<EnvelopeSendInfo>

Is envelope relative to track send\receive\hw_out

None if not.

pub fn automation_state(&self) -> EnvelopeAutomationFlags

current envelope automation state

pub fn tcp_y_offset(&self) -> usize

Y offset of envelope relative to parent track

(may be separate lane or overlap with track contents)

pub fn tcp_y_offset_wo_padding(&self) -> usize

Y offset of envelope relative to parent track, exclusive of padding

pub fn tcp_height(&self) -> usize

visible height of envelope

pub fn tcp_height_wo_padding(&self) -> usize

visible height of envelope, exclusive of padding

impl<'a, P: KnowsProject> Envelope<'a, P, Mutable>

pub fn add_automation_item(
&'a mut self,
pool_id: usize,
position: Position,
length: impl GetLength
) -> AutomationItem<'a, P, Mutable>

Insert new automation item.

if pool_id > 0 automation item will be a new instance of that pool (which will be created as an empty instance if it does not exist

Otherwise, All underlying points will be collected.

pub fn set_point(
&mut self,
point_index: usize,
position: Option<Position>,
point: EnvelopePoint,
sort: bool
) -> ReaperStaticResult<()>

Probably, if position == None — it can lead to moving point to the beginning

pub fn insert_point(
&mut self,
position: Position,
point: EnvelopePoint,
sort: bool
) -> ReaperStaticResult<()>

pub fn delete_point(&mut self, index: usize) -> ReaperStaticResult<()>

pub fn delete_point_range(
&mut self,
start: impl Into<Position>,
end: impl Into<Position>
) -> ReaperStaticResult<()>

pub fn sort_points(&mut self)

pub fn set_state_chunk(
&mut self,
state: impl Into<String>,
with_undo: bool
) -> ReaperStaticResult<()>

Full Envelope state, as it written in project file.