Struct rea_rs::item::Item

pub struct Item<'a, T: ProbablyMutable> { /* private fields */ }

Implementations

impl<'a, T: ProbablyMutable> Item<'a, T>

pub fn new(project: &'a Project, ptr: MediaItem) -> Self

pub fn track(&self) -> Track<'_, Immutable>

pub fn get_take(&'a self, index: usize) -> Option<Take<'a, Immutable>>

pub fn active_take(&'a self) -> Take<'a, Immutable>

pub fn is_selected(&self) -> bool

pub fn position(&self) -> Position

pub fn length(&self) -> Duration

pub fn end_position(&self) -> Position

pub fn is_muted(&self) -> bool

pub fn mute_actual(&self) -> bool

muted (ignores solo). setting this value will not affect Item::solo_override.

pub fn solo_override(&self) -> ItemSoloOverride

Basically, a way to solo particular item.

This function will not override the same parameters on other items

pub fn is_looped(&self) -> bool

pub fn all_takes_play(&self) -> bool

pub fn time_base(&self) -> TimeMode

pub fn auto_stretch(&self) -> bool

pub fn locked(&self) -> bool

pub fn volume(&self) -> Volume

pub fn snap_offset(&self) -> Duration

pub fn fade_in(&self) -> ItemFade

pub fn fade_out(&self) -> ItemFade

pub fn group_id(&self) -> usize

pub fn y_pos_relative(&self) -> usize

Y-position (relative to top of track) in pixels

pub fn y_pos_free_mode(&self) -> f64

Y-position (relative to top of track) in pixels when track is in free mode

0 → top of track, 1 → bottom of track (will never be 1)

pub fn color(&self) -> Option<Color>

if None → default.

pub fn height_free_mode(&self) -> f64

Height in pixels, when track is in free mode

0 → no height (will never be), 1 → full track.

pub fn height(&self) -> usize

height in pixels

pub fn n_takes(&self) -> usize

pub fn notes(
&self,
buf_size: impl Into<Option<u32>>
) -> ReaperStaticResult<String>

pub fn guid(&self) -> GUID

impl<'a> Item<'a, Mutable>

pub fn add_take(&mut self) -> Take<'_, Mutable>

pub fn get_take_mut(&'a mut self, index: usize) -> Option<Take<'a, Mutable>>

pub fn active_take_mut(&'a mut self) -> Take<'a, Mutable>

pub fn set_position(&mut self, position: impl Into<Position>)

pub fn set_length(&mut self, length: Duration)

pub fn set_end_position(&mut self, end_position: impl Into<Position>)

pub fn set_selected(&self, selected: bool)

pub fn delete(self)

pub fn make_only_selected_item(&mut self)

pub fn split(
self,
position: impl Into<Position>
) -> ReaperStaticResult<ItemSplit<'a>>

pub fn update(&mut self)

pub fn move_to_track(&self, track_index: usize) -> ReaperStaticResult<()>

pub fn set_muted(&mut self, muted: bool)

pub fn set_mute_actual(&mut self, mute: bool)

muted (ignores solo). setting this value will not affect Item::solo_override.

pub fn set_solo_override(&mut self, solo: ItemSoloOverride)

Basically, a way to solo particular item.

This function will not override the same parameters on other items

pub fn set_looped(&mut self, looped: bool)

pub fn set_all_takes_play(&mut self, should_play: bool)

pub fn set_time_base(&mut self, mode: TimeMode)

pub fn set_auto_stretch(&mut self, state: bool)

pub fn set_locked(&mut self, locked: bool)

pub fn set_volume(&mut self, volume: Volume)

pub fn set_snap_offset(&mut self, offset: Duration) -> ReaperStaticResult<()>

pub fn set_fade_in(&mut self, fade: ItemFade) -> ReaperStaticResult<()>

pub fn set_fade_out(&mut self, fade: ItemFade) -> ReaperStaticResult<()>

pub fn set_group_id(&mut self, id: usize) -> ReaperStaticResult<()>

pub fn set_y_pos_free_mode(&mut self, y_pos: f64) -> ReaperStaticResult<()>

Y-position (relative to top of track) in pixels when track is in free mode

0 → top of track, 1 → bottom of track (will never be 1)

pub fn set_height_free_mode(&mut self, height: f64)

Height in pixels, when track is in free mode

0 → no height (will never be), 1 → full track.

pub fn set_color(&mut self, color: Option<Color>)

if None → default.

pub fn set_notes(&self, notes: impl Into<String>) -> ReaperStaticResult<()>

pub fn set_guid(&self, guid: GUID) -> ReaperStaticResult<()>