Module rea_rs::misc_types

Structs

Color

DimensionsPixel

GUID

HardwareSocket

Represents Audio\MIDI physical out pin.

Immutable

Guarantees, that object is immutable.

Mutable

Guarantees, that only this object and its child (and sub_child) can be mutated.

Pan

Pitch

PlayRate

Project playback rate.

Position

Position in project.

PositionPixel

RectPixel

SampleAmount

Good helper to be sample-accurate.

TimeRange

TimeSignature

Straightforward TimeSignature, that can be used as Project parameter.

Volume

Enums

PanLaw

PanLawMode

TimeRangeKind

Traits

GetLength

KnowsProject

ProbablyMutable

Generic mutability marker, that allows to mutate only one Reaper object at time.