Enum rea_rs::project::ProjectContext

pub enum ProjectContext {
    CurrentProject,
    Proj(NonNull<ReaProject>),
}

Determines the project in which a function should be executed.

Variants

CurrentProject

Project in the currently open tab.

Proj(NonNull<ReaProject>)

A particular project, not necessarily the one in the currently open tab.

Implementations

impl ProjectContext

pub fn to_raw(self) -> *mut ReaProject

Converts this value to a raw pointer as expected by the low-level API.