Struct rea_rs::reaper::Reaper

pub struct Reaper {
    pub control_surfaces: Vec<Box<dyn ControlSurface>>,
    /* private fields */
}

Fields
control_surfaces: Vec<Box<dyn ControlSurface>>
Implementations

impl Reaper

pub fn make_available_globally(reaper: Reaper)

Makes the given instance available globally.

After this has been called, the instance can be queried globally using get().

This can be called once only. Subsequent calls won’t have any effect!

pub fn load(context: PluginContext)

pub fn low(&self) -> &Reaper

pub fn medium_session(&self) -> &ReaperSession

pub fn medium_session_mut(&mut self) -> &mut ReaperSession

pub fn medium(&self) -> &Reaper

pub fn get() -> &'static Reaper

Gives access to the instance which you made available globally before.

Panics

This panics if make_available_globally() has not been called before.

pub fn get_mut() -> &'static mut Reaper

pub fn register_action(
&mut self,
command_name: &'static str,
description: &'static str,
operation: impl Fn(i32) -> Result<(), Box<dyn Error>> + 'static,
kind: ActionKind
) -> Result<RegisteredAction, Box<dyn Error>>

impl Reaper

pub fn show_console_msg(&self, msg: impl Into<String>)

Show message in console.

Note

\n will be added in the end.

pub fn clear_console(&self)

pub fn perform_action(
&self,
action_id: impl Into<CommandId>,
flag: i32,
project: Option<&Project>
)

Run action by it’s command id.

Note

It seems, that flag should be always 0. If project is None — will perform on current project.

pub fn current_project(&self) -> Project

Get project from the current tab.

pub fn add_project_tab(&self, make_current_project: bool) -> Project

Open new project tab.

To open project in new tab use Reaper::open_project

pub fn open_project(
&self,
file: &Path,
in_new_tab: bool,
make_current_project: bool
) -> Result<Project, &str>

Open project from the filename.

pub fn add_reascript(
&self,
file: &Path,
section: Section,
commit: bool
) -> Result<CommandId, Box<dyn Error>>

Add reascript from file and put to the action list.

commit must be used in the last call, but it is faster to make it false in a bulk.

pub fn remove_reascript(
&self,
file: &Path,
section: Section,
commit: bool
) -> Result<(), Box<dyn Error>>

Remove reascript.

commit must be used in the last call, but it is faster to make it false in a bulk.

pub fn browse_for_file(
&self,
window_title: impl Into<String>,
extension: impl Into<String>
) -> Result<Box<Path>, Box<dyn Error>>

Ask user to select a file.

extension — extension for file, e.g. “mp3”, “txt”. Or empty string.

pub fn arm_command(&self, command: CommandId, section: impl Into<String>)

Arm or disarm command.

Original doc

arms a command (or disarms if 0 passed) in section (empty string for main)

pub fn disarm_command(&self)

pub fn armed_command(&self) -> Option<(CommandId, String)>

Get armed command.

If string is empty (len() = 0), then it’s main section.

pub fn clear_peak_cache(&self)

Reset global peak cache.

pub fn get_action_id(&self, action_name: impl Into<String>) -> Option<CommandId>

Get ID for action with the given name.

Note

name is the ID string, that was made, when registered as action, but not the description line.

If action name doesn’t start with underscore, it will be added.

pub fn get_action_name(&self, id: CommandId) -> Option<String>

Get action name (string ID) of an action with the given ID.

pub fn get_binary_directory(&self) -> String

Return REAPER bin directory (e.g. “C:\Program Files\REAPER”).

pub fn get_global_automation_mode(&self) -> Option<AutomationMode>

Get globally overrided automation mode.

None if do not overrides.

pub fn set_global_automation_mode(&self, mode: AutomationMode)

Override global automation mode.

pub fn get_user_inputs<'a>(
&self,
title: impl Into<String>,
captions: Vec<&'a str>,
buf_size: impl Into<Option<usize>>
) -> ReaperResult<HashMap<String, String>>

Show text inputs to user and get values from them.

Note

default buf size is 1024

pub fn with_prevent_ui_refresh(&self, f: impl Fn())

Call function while freezing the UI.

pub fn with_undo_block(
&self,
undo_name: impl Into<String>,
flags: UndoFlags,
project: Option<&Project>,
f: impl FnMut() -> ReaperResult<()>
) -> ReaperResult<()>

Call function in undo block with given name.

Note

Probably, it’s better to use UndoFlags.all() by default.

pub fn show_message_box(
&self,
title: impl Into<String>,
text: impl Into<String>,
box_type: MessageBoxType
) -> ReaperResult<MessageBoxValue>

Show message box to user and get result.

pub fn update_arrange(&self)

Redraw the arrange view.

pub fn update_timeline(&self)

Redraw timeline.

pub fn view_prefs(
&self,
page: impl Into<Option<u32>>,
name: impl Into<Option<String>>
)

Open preferences window.

page should be positive or None.

if not page — then name will be used.

pub fn iter_projects<'a>(&self) -> ProjectIterator

Iter through all opened projects.

Warning

This operation, probably, of O(n²) complexity. So, it’s better not to use it in loop or too often.

pub fn validate_ptr<'a>(&self, pointer: impl Into<ReaperPointer<'a>>) -> bool

Checks if the given pointer is still valid.

Returns true if the pointer is a valid object of the correct type in the current project.

pub fn validate_ptr_2<'a>(
&self,
project: &Project,
pointer: impl Into<ReaperPointer<'a>>
) -> bool

Checks if the given pointer is still valid.

Example

use reaper_medium::ProjectContext::CurrentProject;

let track = session.reaper().get_track(CurrentProject, 0)
    .ok_or("No track")?;
let track_is_valid = session.reaper()
    .validate_ptr_2(CurrentProject, track);
assert!(track_is_valid);

Returns true if the pointer is a valid object of the correct type in the given project. The project is ignored if the pointer itself is a project.

impl Reaper

pub fn get_latency(&self) -> (SampleAmount, SampleAmount)

Get latency in samples.

Returns (input, output) latency.

pub fn get_approximate_samplerate(&self) -> u32

Try to evaluate samplerate from the latency parameters.

Not stable, and can be not precise.

pub fn audio_init(&self)

Open all audio and MIDI devices (if not opened).

pub fn midi_reinit(&self)

Reset all MIDI devices.

pub fn audio_is_pre_buffer(&self) -> bool

Return whether audio is in pre-buffer (thread safe).

pub fn audio_is_running(&self) -> bool

Return whether audio is running (thread safe).

pub fn get_n_audio_inputs(&self) -> usize

pub fn get_n_audio_outputs(&self) -> usize

pub fn iter_audio_inputs(&self) -> AudioInputsIterator

pub fn iter_audio_outputs(&self) -> AudioOutputsIterator

pub fn get_max_midi_inputs(&self) -> usize

pub fn get_midi_input(&self, index: usize) -> Option<HardwareSocket>

pub fn get_max_midi_outputs(&self) -> usize

pub fn get_midi_output(&self, index: usize) -> Option<HardwareSocket>

pub fn get_audio_input(&self, index: usize) -> Option<HardwareSocket>

pub fn get_audio_output(&self, index: usize) -> Option<HardwareSocket>

pub fn get_n_midi_inputs(&self) -> usize

pub fn get_n_midi_outputs(&self) -> usize

pub fn iter_midi_inputs(&self) -> MidiInputsIterator

pub fn iter_midi_outputs(&self) -> MidiOutputsIterator