Trait rea_rs::send::GenericSendMut

pub trait GenericSendMut<'a>: SendIntType + GenericSend<'a, Mutable> {

    fn delete(self) -> ReaperResult<()>
    where
        Self: Sized,
    { ... }
    fn set_info_value(
        &mut self,
        param: impl Into<String>,
        value: f64
    ) -> ReaperResult<()> { ... }
    fn set_mute(&mut self, mute: bool) -> ReaperResult<()> { ... }
    fn set_phase(&mut self, phase: bool) -> ReaperResult<()> { ... }
    fn set_mono(&mut self, mono: bool) -> ReaperResult<()> { ... }
    fn set_volume(&mut self, volume: impl Into<Volume>) -> ReaperResult<()> { ... }
    fn set_pan(&mut self, pan: impl Into<Pan>) -> ReaperResult<()> { ... }
    fn set_pan_law(&mut self, pan_law: PanLaw) -> ReaperResult<()> { ... }
    fn set_send_mode(&mut self, send_mode: SendMode) -> ReaperResult<()> { ... }
    fn set_automation_mode(
        &mut self,
        automation_mode: AutomationMode
    ) -> ReaperResult<()> { ... }
    fn set_source_channels(
        &mut self,
        channels: Option<SendSourceChannels>
    ) -> ReaperResult<()> { ... }
    fn set_dest_channels(
        &mut self,
        channels: SendDestChannels
    ) -> ReaperResult<()> { ... }
    fn set_midi_properties(
        &mut self,
        properties: impl Into<Option<SendMIDIProps>>
    ) -> ReaperResult<()> { ... }

}

Provided Methods

fn delete(self) -> ReaperResult<()>where
Self: Sized,

Remove send from track. This also drops the value.

Note

drop(send) will not remove send from track.

fn set_info_value(
&mut self,
param: impl Into<String>,
value: f64
) -> ReaperResult<()>

Core method to set send properties. With probability of 99% you shouldn’t use it.

fn set_mute(&mut self, mute: bool) -> ReaperResult<()>

fn set_phase(&mut self, phase: bool) -> ReaperResult<()>

Phase flipped if true.

fn set_mono(&mut self, mono: bool) -> ReaperResult<()>

fn set_volume(&mut self, volume: impl Into<Volume>) -> ReaperResult<()>

fn set_pan(&mut self, pan: impl Into<Pan>) -> ReaperResult<()>

fn set_pan_law(&mut self, pan_law: PanLaw) -> ReaperResult<()>

fn set_send_mode(&mut self, send_mode: SendMode) -> ReaperResult<()>

fn set_automation_mode(
&mut self,
automation_mode: AutomationMode
) -> ReaperResult<()>

Can not set Bypass Use None for default track mode.

fn set_source_channels(
&mut self,
channels: Option<SendSourceChannels>
) -> ReaperResult<()>

Pass None if want to turn audio off.

fn set_dest_channels(&mut self, channels: SendDestChannels) -> ReaperResult<()>

If source channels are off it will return ReaperError::InvalidObject

fn set_midi_properties(
&mut self,
properties: impl Into<Option<SendMIDIProps>>
) -> ReaperResult<()>

Pass None if want to turn MIDI off.