Struct rea_rs::take::Take

pub struct Take<'a, T: ProbablyMutable> { /* private fields */ }

Implementations

impl<'a, T: ProbablyMutable> Take<'a, T>

pub fn new(ptr: MediaItemTake, item: &'a Item<'a, T>) -> Self

pub fn item(&self) -> &Item<'_, T>

pub fn get_visible_fx(&'a self) -> Option<TakeFX<'a, Immutable>>

pub fn is_active(&self) -> bool

pub fn is_midi(&self) -> bool

pub fn n_envelopes(&self) -> usize

pub fn n_fx(&self) -> usize

pub fn n_midi_events(&self) -> usize

pub fn name(&self) -> String

pub fn source(&self) -> Option<Source<'_, Immutable>>

pub fn iter_midi(
&self,
buf_size_override: impl Into<Option<i32>>
) -> ReaperStaticResult<MidiEventBuilder>

Get iterator on human-readable MIDI events.

See crate::midi

pub fn get_midi(
&self,
buf_size_override: impl Into<Option<i32>>
) -> ReaperStaticResult<Vec<u8>>

Get take raw midi data.

It is quite useless as it is, but, it can be used several times with MidiEventBuilder for iterating through various event types.

In the case one desire to iter through raw binary data

MIDI buffer is returned as a list of { int offset, char flag, int msglen, unsigned char msg[] }.

offset: MIDI ticks from previous event

flag: &1=selected &2=muted

flag high 4 bits for CC shape: &16=linear, &32=slow start/end, &16|32=fast start, &64=fast end, &64|16=bezier

msg: the MIDI message.

A meta-event of type 0xF followed by ’CCBZ ’ and 5 more bytes represents bezier curve data for the previous MIDI event: 1 byte for the bezier type (usually 0) and 4 bytes for the bezier tension as a float.

For tick intervals longer than a 32 bit word can represent, zero-length meta events may be placed between valid events.

pub fn guid(&self) -> GUID

pub fn start_offset(&self) -> Duration

pub fn volume(&self) -> Volume

pub fn pan(&self) -> Pan

pub fn pan_law(&self) -> PanLaw

pub fn play_rate(&self) -> PlayRate

pub fn pitch(&self) -> Pitch

take pitch adjustment in semitones, -12=one octave down, 0=normal, +12=one octave up, etc

pub fn preserve_pitch(&self) -> bool

preserve pitch when changing playback rate

pub fn y_pos(&self) -> usize

Y-position (relative to top of track) in pixels (read-only)

pub fn height(&self) -> usize

height in pixels (read-only)

pub fn channel_mode(&self) -> TakeChannelMode

pub fn pitch_mode(&self) -> Option<TakePitchMode>

pub fn color(&self) -> Option<Color>

if None → default.

impl<'a> Take<'a, Mutable>

pub fn add_audio_accessor(
&mut self
) -> AudioAccessor<'a, Take<'_, Mutable>, Mutable>

pub fn set_active(&mut self)

pub fn add_fx(
&mut self,
name: impl Into<String>,
position: impl Into<Option<u8>>,
even_if_exists: bool
) -> Option<TakeFX<'_, Mutable>>

Add FX at given position, or return existing one.

If even_if_exists is false, plugin will be added only if no plugin exists on track.

Otherwise, if position is None → the last slot will be used. The resulting FX will have real index, that may differ from the desired.

pub fn get_fx_mut(&'a mut self, index: usize) -> Option<TakeFX<'a, Mutable>>

pub fn get_visible_fx_mut(&'a mut self) -> Option<TakeFX<'a, Mutable>>

pub fn select_all_midi_events(&mut self, select: bool)

pub fn sort_midi(&mut self)

pub fn source_mut(&mut self) -> Option<Source<'_, Mutable>>

pub fn set_source<T: ProbablyMutable>(
&mut self,
source: Source<'_, T>
) -> ReaperStaticResult<()>

pub fn set_midi(&self, midi: Vec<u8>) -> ReaperStaticResult<()>

Set raw MIDI to take.

Probably, it’s bad idea to construct it DIY, so, see:

crate::midi

Take::get_midi

Take::iter_midi

pub fn set_name(&mut self, name: impl Into<String>)

pub fn set_guid(&mut self, guid: GUID)

pub fn set_start_offset(&mut self, offset: Duration) -> ReaperStaticResult<()>

pub fn set_volume(&mut self, volume: Volume)

pub fn set_pan(&mut self, pan: Pan)

pub fn set_pan_law(&mut self, pan_law: PanLaw)

pub fn set_play_rate(&mut self, play_rate: PlayRate) -> ReaperStaticResult<()>

pub fn set_pitch(&mut self, pitch: Pitch) -> ReaperStaticResult<()>

take pitch adjustment in semitones, -12=one octave down, 0=normal, +12=one octave up, etc

pub fn set_preserve_pitch(&mut self, preserve: bool)

preserve pitch when changing playback rate

pub fn set_channel_mode(&mut self, mode: TakeChannelMode)

pub fn set_pitch_mode(&mut self, mode: Option<TakePitchMode>)

pub fn set_color(&mut self, color: Option<Color>)

if None → default.